merubah method update pada module item, expense, transfer money, receive money, send money

+++ TabpediaFin/TabpediaFin/Handler/ExpenseHandler/ExpenseInsertHandler.cs
@@ -148,10 +148,10 @@ public class ExpenseInsertDto : IRequest<RowResponse<ExpenseFetchDto>>
     public int ContactId { get; set; } = 0;
     public int PaymentMethodId { get; set; } = 0;
     public int PaymentTermId { get; set; } = 0;
-    public Int64 Amount { get; set; } = 0;
+    public int Amount { get; set; } = 0;
     public int DiscountTypeId { get; set; } = 0;
     public int DiscountPercent { get; set; } = 0;
-    public Int64 DiscountAmount { get; set; } = 0;
+    public int DiscountAmount { get; set; } = 0;
     public string Notes { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public int TaxId { get; set; } = 0;
